Cria superclasse abstrata Match para SingleMatch e TagMatch
