+++ ww3/src/main/java/br/edu/insper/desagil/aps5/ww3/SimpleMatch.java
@@ -1,16 +1,12 @@
 package br.edu.insper.desagil.aps5.ww3;
 
-public class SimpleMatch {
+public class SimpleMatch extends Match {
     private Wrestler faceWrestler;
     private Wrestler heelWrestler;
-    private boolean faceAdvantage;
-    private boolean heelAdvantage;
 
     public SimpleMatch(Wrestler faceWrestler, Wrestler heelWrestler) {
         this.faceWrestler = faceWrestler;
         this.heelWrestler = heelWrestler;
-        this.faceAdvantage = false;
-        this.heelAdvantage = false;
     }
 
     public Wrestler getFaceWrestler() {
@@ -21,25 +17,9 @@ public class SimpleMatch {
         return heelWrestler;
     }
 
-    public boolean isFaceAdvantage() {
-        return faceAdvantage;
-    }
-
-    public void setFaceAdvantage(boolean faceAdvantage) {
-        this.faceAdvantage = faceAdvantage;
-    }
-
-    public boolean isHeelAdvantage() {
-        return heelAdvantage;
-    }
-
-    public void setHeelAdvantage(boolean heelAdvantage) {
-        this.heelAdvantage = heelAdvantage;
-    }
-
     public void finish() {
-        if (faceAdvantage) {
-            if (heelAdvantage) {
+        if (isFaceAdvantage()) {
+            if (isHeelAdvantage()) {
                 faceWrestler.addPoints(1);
                 heelWrestler.addPoints(1);
             } else {
@@ -47,7 +27,7 @@ public class SimpleMatch {
                 heelWrestler.addPoints(0);
             }
         } else {
-            if (heelAdvantage) {
+            if (isHeelAdvantage()) {
                 faceWrestler.addPoints(0);
                 heelWrestler.addPoints(2);
             } else {
