+++ ww3/src/main/java/br/edu/insper/desagil/aps5/ww3/TagMatch.java
@@ -1,16 +1,12 @@
 package br.edu.insper.desagil.aps5.ww3;
 
-public class TagMatch {
+public class TagMatch extends Match {
     private Team faceTeam;
     private Team heelTeam;
-    private boolean faceAdvantage;
-    private boolean heelAdvantage;
 
     public TagMatch(Team faceTeam, Team heelTeam) {
         this.faceTeam = faceTeam;
         this.heelTeam = heelTeam;
-        this.faceAdvantage = false;
-        this.heelAdvantage = false;
     }
 
     public Team getFaceTeam() {
@@ -21,25 +17,9 @@ public class TagMatch {
         return heelTeam;
     }
 
-    public boolean isFaceAdvantage() {
-        return faceAdvantage;
-    }
-
-    public void setFaceAdvantage(boolean faceAdvantage) {
-        this.faceAdvantage = faceAdvantage;
-    }
-
-    public boolean isHeelAdvantage() {
-        return heelAdvantage;
-    }
-
-    public void setHeelAdvantage(boolean heelAdvantage) {
-        this.heelAdvantage = heelAdvantage;
-    }
-
     public void finish() {
-        if (faceAdvantage) {
-            if (heelAdvantage) {
+        if (isFaceAdvantage()) {
+            if (isHeelAdvantage()) {
                 faceTeam.getFirst().addPoints(0.5);
                 faceTeam.getSecond().addPoints(0.5);
                 heelTeam.getFirst().addPoints(0.5);
@@ -51,7 +31,7 @@ public class TagMatch {
                 heelTeam.getSecond().addPoints(0);
             }
         } else {
-            if (heelAdvantage) {
+            if (isHeelAdvantage()) {
                 faceTeam.getFirst().addPoints(0);
                 faceTeam.getSecond().addPoints(0);
                 heelTeam.getFirst().addPoints(1);
